refactor: update type imports after AdminPanel->admin rename (#55)
